refactor: break down TechnicalDetailsModal into modular components
- Refactor monolithic 1471-line TechnicalDetailsModal into 8 focused components
- Extract sections: ModalHeader, TabNavigation, ScreenshotGallery, and 6 section cards
- Create comprehensive TypeScript types with ExtendedTechnicalSection interface
- Implement proper type safety, removing all 'any' type warnings
- Add utility functions for icon mapping and section visibility logic
- Update dynamic imports to handle new component structure
- Maintain all existing functionality while improving maintainability
- Achieve zero ESLint warnings and successful build compilation

Components created:
- ModalHeader: Project title, description, and dynamic icon display
- TabNavigation: Interactive tab switching with hover effects
- ScreenshotGallery: Image modal with zoom functionality
- Section Cards: OverviewCard, MetricsCard, ToolsCard, MonitoringCard,
  WorkflowsCard, PerformanceCard, ArchitectureCard, DeploymentCard
- Utils: getTechnicalIcon, shouldShowSection, formatSectionTitle
- Types: Comprehensive TypeScript interfaces for all component props

This refactoring improves code organization, reusability, and maintainability
while preserving the exact same user experience and functionality.
